fix: Updating examples for AppDaemon 3.0
As of AppDaemon 3.0, namespaces have been introduced to make it
possible to handle multiple instances of the same plugin.  These
changes bring the upstream in-line with the version from the
examples repository.

This change should be congruent with the changes in
home-assistant/appdaemon@6f97b5927e3b7b64902d474a3e8aed00024612f8